Add an inactivity timeout to the Presto UI
For FedRamp and HIPAA compliance, we wish to limit
user sessions. A new config property
(default of 0, disabled) will control the timeout
duration

A new  is added which the UI app will
redirect to after the timeout duration. The redirect
only happens if the user is observed to be inactive
on the webpage. A call to new endpoint
is made on page load, and this invalidates the basic
auth token for the default realm, thereby requiring
the user to re-login into the UI

Caveat -
Because of how browsers implement basic auth, a re-login
from the  page will *not* work. The user must
cancel any login windows and re-open the  app for a
new, successful auth flow

Co-authored-by: Li Zhou <lizhou@ibm.com>

Alchemy-item: (ID = 903) [Forward-fit] Add an inactivity timeout to the Presto UI commit 1/1 - 7fa3c178041d6596a2f7168b4c34d8129d663765